Move execution witness RLP encoding tests from test_partial_trie_db.cpp into a separate test_execution_witness.cpp file
